fixed all routed between pages
