+++ src/App.js
@@ -4,14 +4,13 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Header from "./components/header/header";
 import Page404 from "./components/error/Page404";
 import Footer from "./components/footer/Footer";
+import Game from "./components/Game/Game";
+import LandingPage from "./components/LandingPage/LandingPage";
+import Register from "./components/Register/Register";
+import Login from "./components/Login/Login";
 
 
 
-// function Game(){
-//   return(
-//     <div>Game</div>
-//   )
-// }
 class App extends Component {
   render() {
     return (
