+++ src/components/header/header.js
@@ -7,9 +7,9 @@ function Header() {
       <Navbar bg="light" variant="light">
           <Navbar.Brand href="#landingPage"></Navbar.Brand>
           <Nav className="mr-auto">
-            <Nav.Link href="#Register">Sign Up</Nav.Link>
-            <Nav.Link href="#Login">Login</Nav.Link>
-            <Nav.Link href="#Game">Play Now</Nav.Link>
+            <Nav.Link href="Register">Sign Up</Nav.Link>
+            <Nav.Link href="Login">Login</Nav.Link>
+            <Nav.Link href="Game">Play Now</Nav.Link>
           </Nav>
           </Navbar>
           )
