+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 //import App from './App';
-import Game from './Game';
+import Game from './components/Game/Game';
 import './index.css';
 import {Provider} from 'react-redux';
 import store from './components/config/store';
